Math Context into Funstion Strategies. change name of the interface
