+++ src/test/java/com/github/bgora/rpnlibrary/FibTest.java
@@ -31,7 +31,7 @@ import static org.junit.Assert.assertEquals;
 
 public class FibTest {
 
-    private CalculatorInterface calc;
+    private Calculator calc;
 
     @Before
     public void setUp() throws Exception {
